+++ infrastructure/tencentcloud/.terraform.lock.hcl
@@ -1,0 +1,21 @@
+# This file is maintained automatically by "terraform init".
+# Manual edits may be lost in future updates.
+
+provider "registry.terraform.io/tencentcloudstack/tencentcloud" {
+  version     = "1.53.7"
+  constraints = "1.53.7"
+  hashes = [
+    "h1:ettmtC3q3wcby+eQ9WvQdF7YGsbtax8+T4t7OjVVr/M=",
+    "zh:39c41219152bdab3f19c318fbe12b6c6587567279480e069bcdfebb090b6ec67",
+    "zh:42f48d2400927359e9be2895e7b78c2dfb8ffc4c11e3ece435a85128791e3720",
+    "zh:46135ff3688545cb66b50fbeb8c988a4060585451a5ee9ec7a15317c6d4ec911",
+    "zh:47bc50fa8e53a73f32ef43628604ef0e42266208abf9503c36413c1750a1109e",
+    "zh:4e607240c1e2239aa803d84c9932666da19a44a2c94007e5094dda79c9d75621",
+    "zh:5b4cd689a2c0f62261c023be127f61ff0133b372921464c887064e9063837968",
+    "zh:670e846488a609b1173890f375adfef3487caa68610d8550f7272544a24d4bf3",
+    "zh:a72cb73d872438f6479f28a31227cef903813a3a0e74b833a4b8711e93724eed",
+    "zh:be72aaca155ba7bc4f7786219d764bc04b5ac0c10e575be162904b4ae4b6459b",
+    "zh:d1a6e953f06c02de9317507430fd9f13f8311af0bc955ff26b749272325af3d8",
+    "zh:fc2c2fa8b0242989f51bdf0f43df09e9afa9e497bba874d96a9ec4a8e16d9ed4",
+  ]
+}
