+++ module/phrase-generator/Dockerfile
@@ -1,0 +1,10 @@
+FROM golang:latest as builder
+WORKDIR /build
+COPY . /build
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o phrase
+
+FROM alpine:latest
+WORKDIR /app
+COPY --from=builder /build/phrase .
+EXPOSE 9090
+CMD ["/app/phrase"]
