+++ module/phrase-generator/go.mod
@@ -1,0 +1,5 @@
+module phrase-generator
+
+go 1.16
+
+require github.com/labstack/echo/v4 v4.2.1 // indirect
